Add Sugar ORM and Post table
The Post object represents a db object (managed by Sugar ORM)
EditActivity can now be created with a post id. If one is not supplied it
will create a new Post object.
The save button will now save the post to the database.
Currently there is no way to load a saved post. That is next